[BE] - Reorganization
